Add packaging scripts
Rename module name from "src" to "mcnn" which sounds like a
package name not just a directory name.

Add setup.py to make it a package, then it can be install by pip.

All dependencies are listed in requirements.txt, however did not
give specific version constraint. It can be refine in future.